+++ util/ldb_cmd.cc
@@ -281,6 +281,8 @@ Options LDBCommand::PrepareOptionsForOpenDB() {
       opt.compression = kLZ4Compression;
     } else if (comp == "lz4hc") {
       opt.compression = kLZ4HCCompression;
+    } else if (comp == "zstd") {
+      opt.compression = kZSTDNotFinalCompression;
     } else {
       // Unknown compression.
       exec_state_ =